requirements.txt
# Basic requirements file for Scientia project

# Core dependencies
numpy>=1.20.0
faiss-cpu>=1.7.3
requests==2.31.0
openai==1.12.0
firecrawl==0.1.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
nltk>=3.8.0

# Document processing
langchain>=0.0.267
markdown>=3.4.0

# Database and storage
sqlalchemy>=1.4.0
# sqlite3>=0.0.0  # Removed: this is part of Python's standard library
chromadb>=0.3.0  # Alternative vector DB

# Web scraping
html2text>=2020.1.16
trafilatura>=1.0.0
lxml>=4.9.0

# Utilities
tqdm>=4.62.0
python-dotenv==1.0.1
pyyaml>=6.0

# Testing (optional)
pytest>=7.0.0